+++ R/light_check.R
@@ -3,6 +3,7 @@
 #' Checks if an object of class \code{flashlight} or \code{multiflashlight} is consistently defined.
 #'
 #' @param x An object of class \code{flashlight} or \code{multiflashlight}.
+#' @param ... Further arguments passed from or to other methods.
 #' @return The input \code{x} or an error message.
 #' @export
 #' @examples
@@ -12,19 +13,19 @@
 #' fl_log <- flashlight(fit_log,  y = "Sepal.Length", label = "ols", linkinv = exp)
 #' light_check(fl)
 #' light_check(fl_log)
-light_check <- function(x) {
+light_check <- function(x, ...) {
   UseMethod("light_check")
 }
 
 #' @describeIn light_check Default check method not implemented yet.
 #' @export
-light_check.default <- function(x) {
+light_check.default <- function(x, ...) {
   stop("No default method available yet.")
 }
 
 #' @describeIn light_check Checks if a flashlight object is consistently defined.
 #' @export
-light_check.flashlight <- function(x) {
+light_check.flashlight <- function(x, ...) {
   if (is.null(x$label)) {
     stop("label should not be NULL.")
   }
@@ -63,7 +64,7 @@ light_check.flashlight <- function(x) {
 
 #' @describeIn light_check Checks if a multiflashlight object is consistently defined.
 #' @export
-light_check.multiflashlight <- function(x) {
+light_check.multiflashlight <- function(x, ...) {
   # by
   if (!all_identical(x, `[[`, "by")) {
     warning("Inconsistent 'by' variables specified.