+++ R/light_interaction.R
@@ -2,7 +2,7 @@
 #'
 #' This function provides Friedman's H statistic [1] for overall interaction strength per covariable as well as its version for pairwise interactions. As a fast alterantive to assess overall interaction strength, with \code{type = "ice"}, the function offers a method based on centered ICE curves: The corresponding H* statistic measures how much of the variability of a c-ICE curve is unexplained by the main effect. As for Friedman's H statistic, it can be useful to consider unnormalized or squared values (see Details below).
 #'
-#' Friedman's H statistic relates the interaction strength of a variable (pair) to the total effect strength of that variable (pair) based on partial dependence curves. Due to this normalization step, even variables with low importance can have high values for H. The function \code{light_interaction} offers the option to skip this normalization step in order to have a more direct comparison of the interaction effects across variable (pairs). The values of such unnormalized H are on the scale of the response variable. Use \code{take_sqrt = FALSE} to return squared values of H. Note that in general, for each variable (pair) predicitons are done on a data set with \code{grid_size * n_max}, so be cautious with increasing the defaults too much.
+#' Friedman's H statistic relates the interaction strength of a variable (pair) to the total effect strength of that variable (pair) based on partial dependence curves. Due to this normalization step, even variables with low importance can have high values for H. The function \code{light_interaction} offers the option to skip this normalization step in order to have a more direct comparison of the interaction effects across variable (pairs). The values of such unnormalized H are on the scale of the response variable. Use \code{take_sqrt = FALSE} to return squared values of H. Note that in general, for each variable (pair) predicitons are done on a data set with \code{grid_size * n_max}, so be cautious with increasing the defaults too much. The minimum required elements in the (multi-) flashlight are a "predict_function", "model", and "data".
 #'
 #' @importFrom stats setNames
 #' @importFrom dplyr as_tibble bind_rows bind_cols group_by_at do ungroup
@@ -25,15 +25,18 @@
 #' @param label_name Column name in resulting \code{data} containing the label of the flashlight. Defaults to "label".
 #' @param error_name Currently not used.
 #' @param variable_name Column name in resulting \code{data} containing the variable names. Defaults to "variable".
+#' @param type_name Column name in the resulting \code{data} with the plot \code{type}.
 #' @param ... Further arguments passed to or from other methods.
 #' @return An object of class \code{light_importance}, \code{light} (and a list) with the following elements.
 #' \itemize{
 #'   \item \code{data} A tibble containing the results. Can be used to build fully customized visualizations. Its column names are specified by the items in this list (except for "method").
 #'   \item \code{by} Same as input \code{by}.
+#'   \item \code{type} Same as input \code{type}. For information only.
 #'   \item \code{value_name} Same as input \code{value_name}.
 #'   \item \code{error_name} Same as input \code{error_name}.
 #'   \item \code{label_name} Same as input \code{label_name}.
 #'   \item \code{variable_name} Same as input \code{variable_name}.
+#'   \item \code{type_name} Same as input \code{type_name}.
 #' }
 #' @export
 #' @references [1] Friedman, J. H. and Popescu, B. E. (2008). “Predictive learning via rule ensembles.” The Annals of Applied Statistics. JSTOR, 916–54..
@@ -67,7 +70,8 @@ light_interaction.flashlight <- function(x, data = x$data, by = x$by,
                                          seed = NULL, use_linkinv = FALSE,
                                          value_name = "value",
                                          error_name = "error", label_name = "label",
-                                         variable_name = "variable", ...) {
+                                         variable_name = "variable",
+                                         type_name = "type", ...) {
   # Checks
   type <- match.arg(type)
   cols <- colnames(data)
@@ -192,12 +196,14 @@ light_interaction.flashlight <- function(x, data = x$data, by = x$by,
   # Prepare output
   agg[[label_name]] <- x$label
   agg[[error_name]] <- NA
+  agg[[type_name]] <- type
 
   # Collect results
   var_order <- c(label_name, by, variable_name, value_name, error_name)
-  out <- list(data = agg[, var_order], by = by,
+  out <- list(data = agg[, var_order], by = by, type = type,
               value_name = value_name, error_name = error_name,
-              label_name = label_name, variable_name = variable_name)
+              label_name = label_name, variable_name = variable_name,
+              type_name = type_name)
   class(out) <- c("light_importance", "light", "list")
   out
 }